requirements.txt
bmp280==0.0.4
pyserial==3.5
RPi.GPIO==0.7.1
smbus==1.1.post2
smbus2==0.4.3